Delete unused helper modules and reorganize helper methods
