--- app/helpers/bodyweights_helper.rb
@@ -1,2 +1,0 @@
-module BodyweightsHelper
-end
