--- app/helpers/cardios_helper.rb
@@ -1,2 +1,0 @@
-module CardiosHelper
-end
