--- app/helpers/certifications_helper.rb
@@ -1,2 +1,0 @@
-module CertificationsHelper
-end
