--- app/helpers/contests_helper.rb
@@ -1,2 +1,0 @@
-module ContestsHelper
-end
