--- app/helpers/conversations_helper.rb
@@ -1,2 +1,0 @@
-module ConversationsHelper
-end
