--- app/helpers/invitations_helper.rb
@@ -1,2 +1,0 @@
-module InvitationsHelper
-end
