+++ app/helpers/messages_helper.rb
@@ -1,2 +1,7 @@
 module MessagesHelper
+
+  def new_message_count_print
+    new_message_count > 0 ? "(#{new_message_count})" : "" if new_message_count
+  end
+
 end
