--- app/helpers/philosophies_helper.rb
@@ -1,2 +1,0 @@
-module PhilosophiesHelper
-end
