--- app/helpers/photos_helper.rb
@@ -1,2 +1,0 @@
-module PhotosHelper
-end
