--- app/helpers/resourceries_helper.rb
@@ -1,7 +1,0 @@
-module ResourceriesHelper
-
-  def resource_shared_with_prep(resource)
-    resource.preps.include? @prep
-  end
-
-end
