--- app/helpers/target_cardios_helper.rb
@@ -1,2 +1,0 @@
-module TargetCardiosHelper
-end
