--- app/helpers/target_macros_helper.rb
@@ -1,2 +1,0 @@
-module TargetMacrosHelper
-end
